Marcus new/improved sever auth (#27)
* created registration endpoint

* created file for authentication and authorization on server side

* updated README

* updated session to store token

* removed console statements

* sucessfully server auth with cookie

* removed token from session

* removed server consoles

* added method limitation to delete endpoint

* abstracted uid and role retrieval to helper function

+++ src/pages/api/users/delete.ts
@@ -1,5 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { deleteUser } from '@/data/userData';
+import getUIDandRole from '@/data/serverAuth';
 
 export default async function handler(
   req: NextApiRequest,
@@ -9,17 +10,36 @@ export default async function handler(
     return res.status(405).json({ error: 'Method not allowed' });
   }
 
+  let uid: string;
+  let role: string;
+
+  try {
+    const data = await getUIDandRole(req);
+    uid = data.uid;
+    role = data.role;
+  } catch (error) {
+    console.error(error);
+    return res.status(401).json({ error: 'Not authorized' });
+  }
+
   const { id } = req.body;
 
   if (!id) {
     return res.status(400).json({ error: 'ID is required' });
   }
 
+  // Authorization: Check if the requester is deleting their own account or is an admin
+  if (uid !== id && role !== 'admin') {
+    return res.status(403).json({
+      error: 'Forbidden: You do not have permission to perform this action.',
+    });
+  }
+
   try {
     await deleteUser(id);
-    res.status(200).json({ message: 'User deleted successfully' });
+    return res.status(200).json({ message: 'User deleted successfully' });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: 'Failed to delete user' });
+    return res.status(500).json({ error: 'Failed to delete user' });
   }
 }
